+++ Atmel_Smart_Watch_Project/SmartWatch/SmartWatch/src/smartwatch/display_manager.h
@@ -40,6 +40,6 @@
 
 void display_manager_init(void);
 void display_ui_task(uint8_t button);
-void updateGraph(float glucose);
+void updateGraph(uint16_t glucose);
 
 #endif
